requirements.txt
# Core dependencies
transformers>=4.30.0
torch>=2.0.0
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Monitoring and logging
prometheus-client>=0.16.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-prometheus>=1.18.0
python-json-logger>=2.0.7

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
requests>=2.30.0

# AWS SDK (if deploying to AWS)
boto3>=1.26.0

# GCP SDK (if deploying to GCP) 
google-cloud-storage>=2.9.0

# Azure SDK (if deploying to Azure)
azure-storage-blob>=12.16.0
azure-identity>=1.13.0